# e2e/openapi.yml
...
openapi: 3.0.1
...
info:
  title: Conduit API
  description: Conduit API
  contact:
    name: RealWorld
    url: https://realworld.io
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0

#
# OpenAPI Object: servers
#
# 型
# - [Server Object]
# 必須？
# - No
# URL
# - https://spec.openapis.org/oas/v3.0.1#openapi-object
# 概要
# - ターゲットサーバへの接続情報を提供するServer Objectの配列
# Server Objectの詳細
# - https://spec.openapis.org/oas/v3.0.1#server-object
#
servers:
  - url: /api

#
# OpenAPI Object: paths
#
# 型
# - Paths Object
# 必須？
# - Yes
# URL
# - https://spec.openapis.org/oas/v3.0.1#openapi-object
# 概要
# - 利用可能なパスと操作内容
# Paths Objectの詳細
# - https://spec.openapis.org/oas/v3.0.1#paths-object
#
paths:
...
  /articles/feed:
    get:
      tags:
        - Articles
      summary: Get recent articles from users you follow
      description: Get most recent articles from users you follow. Use query parameters
        to limit. Auth is required
      operationId: GetArticlesFeed
      parameters:
        # 自前追加
        - name: Authorization
          in: header
          description: セッション用JWT
          required: true
          schema:
            $ref: '#/components/securitySchemes/Token'
        - name: limit
          in: query
          description: Limit number of articles returned (default is 20)
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Offset/skip number of articles (default is 0)
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipleArticlesResponse'
        401:
          description: Unauthorized
          content: {}
        422:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
      security:
        - Token: []
  /articles:
    get:
      tags:
        - Articles
      summary: Get recent articles globally
      description: Get most recent articles globally. Use query parameters to filter
        results. Auth is optional
      operationId: GetArticles
      parameters:
        #
        # 自前追加
        #
        - name: Authorization
          in: header
          description: セッション用JWT
          required: false
          schema:
            $ref: '#/components/securitySchemes/Token'
        - name: tag
          in: query
          description: Filter by tag
          schema:
            type: string
        - name: author
          in: query
          description: Filter by author (username)
          schema:
            type: string
        - name: favorited
          in: query
          description: Filter by favorites of a user (username)
          schema:
            type: string
        - name: limit
          in: query
          description: Limit number of articles returned (default is 20)
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Offset/skip number of articles (default is 0)
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultipleArticlesResponse'
        401:
          description: Unauthorized
          content: {}
        422:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
    post:
      tags:
        - Articles
      summary: Create an article
      description: Create an article. Auth is required
      operationId: CreateArticle
      # 自前追加
      parameters:
        - name: Authorization
          in: header
          description: セッション用JWT
          required: true
          schema:
            $ref: '#/components/securitySchemes/Token'
      requestBody:
        description: Article to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewArticleRequest'
        required: true
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleArticleResponse'
        401:
          description: Unauthorized
          content: {}
        422:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
      security:
        - Token: []
      x-codegen-request-body-name: article
  /articles/{slug}:
    get:
      tags:
        - Articles
      summary: Get an article
      description: Get an article. Auth not required
      operationId: GetArticle
      parameters:
        - name: slug
          in: path
          description: Slug of the article to get
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleArticleResponse'
        422:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
    put:
      tags:
        - Articles
      summary: Update an article
      description: Update an article. Auth is required
      operationId: UpdateArticle
      parameters:
        # 自前追加
        - name: Authorization
          in: header
          description: セッション用JWT
          required: true
          schema:
            $ref: '#/components/securitySchemes/Token'
        - name: slug
          in: path
          description: Slug of the article to update
          required: true
          schema:
            type: string
      requestBody:
        description: Article to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleArticleResponse'
        401:
          description: Unauthorized
          content: {}
        422:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
      security:
        - Token: []
      x-codegen-request-body-name: article
    delete:
      tags:
        - Articles
      summary: Delete an article
      description: Delete an article. Auth is required
      operationId: DeleteArticle
      parameters:
        # 自前追加
        - name: Authorization
          in: header
          description: セッション用JWT
          required: true
          schema:
            $ref: '#/components/securitySchemes/Token'
        - name: slug
          in: path
          description: Slug of the article to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content: {}
        401:
          description: Unauthorized
          content: {}
        422:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
      security:
        - Token: []
...
components:
...
  schemas:
...
    Profile:
      required:
        - bio
        - following
        - image
        - username
      type: object
      properties:
        username:
          type: string
        bio:
          type: string
        image:
          type: string
        following:
          type: boolean
    Article:
      required:
        - author
        - body
        - createdAt
        - description
        - favorited
        - favoritesCount
        - slug
        - tagList
        - title
        - updatedAt
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        body:
          type: string
        tagList:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        favorited:
          type: boolean
        favoritesCount:
          type: integer
        author:
          $ref: '#/components/schemas/Profile'
    SingleArticleResponse:
      required:
        - article
      type: object
      properties:
        article:
          $ref: '#/components/schemas/Article'
    MultipleArticlesResponse:
      required:
        - articles
        - articlesCount
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        articlesCount:
          type: integer
    NewArticle:
      required:
        - body
        - description
        - title
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        body:
          type: string
        tagList:
          type: array
          items:
            type: string
    NewArticleRequest:
      required:
        - article
      type: object
      properties:
        article:
          $ref: '#/components/schemas/NewArticle'
    UpdateArticle:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        body:
          type: string
    UpdateArticleRequest:
      required:
        - article
      type: object
      properties:
        article:
          $ref: '#/components/schemas/UpdateArticle'
...
    GenericErrorModel:
      required:
        - errors
      type: object
      properties:
        errors:
          required:
            - body
          type: object
          properties:
            body:
              type: array
              items:
                type: string
    #
    # 自前追加
    #
    # '#/components/securitySchemes/Token' をschemeに設定した時
    # Tokenモデルがないため、エラーになる
    # TokenはStringでいいので定義する
    #
...
  securitySchemes:

    #
    # Security Scheme Object
    #
    # 使用可能なセキュリティスキーム定義
    # - HTTP認証
    # - APIキー
    #   - ヘッダ
    #   - クエリパラメータ
    # - OAuth2
    #   - implicit
    #   - password
    #   - application
    #   - access code
    # - OIDC
    #
    Token:
      #
      # Security Scheme Object: type
      #
      # 型
      # - string
      #   - "apiKey"
      #   - "http"
      #   - "oauth2"
      #   - "openIdConnect"
      # 必須？
      # - Yes
      # URL
      # - https://spec.openapis.org/oas/v3.0.1#security-scheme-object
      # 概要
      # - セキュリティスキームのタイプ
      #
      type: apiKey

      #
      # Security Scheme Object: description
      #
      # 型
      # - string
...